+++ index.html
@@ -235,6 +235,5 @@
     </template>
 
 <script src="./js/main.js" type="module"></script>
-<script src="js/functions.js"></script>
 </body>
 </html>
