+++ js/constants.js
@@ -1,0 +1,40 @@
+const MESSAGES = [
+  'Всё отлично!',
+  'В целом всё неплохо. Но не всё.',
+  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
+  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
+  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
+  'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!'
+];
+
+const NAMES = ['Артём', 'Настя', 'Владимир', 'Мария', 'Игорь', 'Светлана', 'Алексей', 'Даша', 'Михаил', 'Оля'];
+
+const DESCRIPTIONS = [
+  'Отличный день',
+  'Хорошее утро',
+  'День идет неплохо'
+];
+
+const AVATAR_MIN = 1;
+const AVATAR_MAX = 6;
+
+const COMMENT_MIN = 0;
+const COMMENT_MAX = 30;
+
+const LIKES_MIN = 15;
+const LIKES_MAX = 200;
+
+const TOTAL_CARDS = 25;
+
+export {
+  MESSAGES,
+  NAMES,
+  DESCRIPTIONS,
+  AVATAR_MIN,
+  AVATAR_MAX,
+  COMMENT_MIN,
+  COMMENT_MAX,
+  LIKES_MIN,
+  LIKES_MAX,
+  TOTAL_CARDS
+}
