+++ js/render.js
@@ -1,0 +1,3 @@
+export const renderCards = (pictures) => {
+    console.log(pictures)
+}
